Added nvim-project and nvim-notify
Added a hook that will look for project.lua file in any project folder
loaded. If this file then it expects the file to return a table with an
init function that takes no arguments and returns nothing. This function
is then called.

You can setup project specific key bindings and other options