Project load will also load exrc file
Unfortunately exrc files are only loaded when vim starts up in a folder with a .nvim.lua file.
If you use the project <leader>p plugin and switch projects, it will not load the nvim.lua file

So I added this to my project settings loader. In addition to loding a project.lua file, if
a .nvim.lua file is found in a project folder, it will be sourced.

This is not very secure becuase anything could be in that file. exrc atleast has a UI to ask
if you want to do this with improvements like only asking once until the file has been modified.
I have no idea how to implement this so i didn't do it. Meaning Insecure. God Help me.