Added some support for busted lua spec files
This will add a key map to run the currently opened busted spec file. This is meant for testing plugin code.
It creates a filetype called busted spec which inherits from the lua file type. It then adds a file type plugin which
loads the custom key map for the file.